update structure of flow
