+++ app/flow/flow_factory.py
@@ -1,8 +1,13 @@
+from enum import Enum
 from typing import Dict, List, Union
 
 from app.agent.base import BaseAgent
-from app.flow.base import BaseFlow, FlowType
+from app.flow.base import BaseFlow
 from app.flow.planning import PlanningFlow
+
+
+class FlowType(str, Enum):
+    PLANNING = "planning"
 
 
 class FlowFactory:
